refactor(at_client_flutter): removed atsign_key.dart and introduced at_client_data.dart
